+++ Control/angle.h
@@ -2,7 +2,7 @@
 #define _angle_h
 
 void angle_correction(void);
-int angle_pid_control(float tar);
-void angleloop_pid_control(float angle_tar, int base);
+int angle_pid_control(int tar);
+void angleloop_pid_control(int angle_tar, int base);
 
 #endif
